purger function for casbotany database now functional
