+++ image_client/csv_purge_sql/image_server_purge.sql
@@ -1,1 +1,1 @@
-DELETE FROM images WHERE datetime  > '2023-07-27 14:56:00.130227' AND datetime < '2023-07-27 14:56:50.121451';
+DELETE FROM images WHERE datetime  > (now() - interval 10 day);
