+++ image_client/csv_purge_sql/picturae_purge_script.sql
@@ -1,8 +1,8 @@
-DELETE FROM casbotany.collectionobjectattachment WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.attachment WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.determination WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.collectionobject WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.collector WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.collectingevent WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.locality WHERE TimestampCreated > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
-DELETE FROM casbotany.agent WHERE TimestampCreated  > '2023-07-27 14:56:00.130227' AND TimestampCreated < '2023-07-27 14:56:31.121451';
+DELETE FROM casbotany.collectionobjectattachment WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.attachment WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.determination WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.collectionobject WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.collector WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.collectingevent WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.locality WHERE TimestampCreated  > (now() - interval 10 day);
+DELETE FROM casbotany.agent WHERE TimestampCreated  > (now() - interval 10 day);
