+++ image_client/csv_purge_sql/upload_time_stamps.csv
@@ -1,0 +1,2 @@
+StartTime,EndTime,UploadCode
+2023-07-28 09:28:13.238155,2023-07-28 09:28:45.936700,476987
